Refactored conversion test to appropirate location in tests/parser/functions directory
