+++ tests/parser/types/test_bytes.py
@@ -4,24 +4,6 @@ from vyper.exceptions import (
 )
 
 
-def test_convert_from_bool(get_contract_with_gas_estimation):
-    exp_code = """
-@public
-def testConvertBytes32(flag: bool) -> bytes32:
-    flagBytes: bytes32 = convert(flag, bytes32)
-    return flagBytes
-    """
-
-    c = get_contract_with_gas_estimation(exp_code)
-    falseBytes = c.testConvertBytes32(False)
-    assert falseBytes[31:32] == b'\x00'
-    assert len(falseBytes) == 32
-
-    trueBytes = c.testConvertBytes32(True)
-    assert trueBytes[31:32] == b'\x01'
-    assert len(trueBytes) == 32
-
-
 def test_address_bytes_converstion(get_contract_with_gas_estimation):
     # Shared values for test addresses and bytes
     test_address = "0xF5D4020dCA6a62bB1efFcC9212AAF3c9819E30D7"
